Implemented optional suggestions from previous code review
