+++ lib/nameable.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name
+    raise NotImplementedError, 'method not implemented yet'
+  end
+end
