+++ lib/person.rb
@@ -1,8 +1,5 @@
-class Nameable
-  def correct_name
-    raise NotImplementedError, 'method not implemented yet'
-  end
-end
+require_relative 'nameable'
+require_relative 'decorators'
 
 class Person < Nameable
   attr_accessor :name, :age
@@ -36,28 +33,5 @@ class Person < Nameable
     return true if @age >= 18
 
     false
-  end
-end
-
-class Decorator < Nameable
-  def initialize(nameable)
-    super()
-    @nameable = nameable
-  end
-
-  def correct_name
-    @nameable.correct_name
-  end
-end
-
-class CapitalizeDecorator < Decorator
-  def correct_name
-    super.capitalize
-  end
-end
-
-class TrimmerDecorator < Decorator
-  def correct_name
-    super.slice(0, 10)
   end
 end
